requirements.txt
apify-client
python-dotenv
pandas
requests
sentence-transformers
scikit-learn
python-docx
anthropic
openpyxl
